Move the logger into core.
