+++ build.py
@@ -82,7 +82,6 @@ cxxlibrary(
         "./lib/imagewriter/rawimagewriter.cc",
         "./lib/layout.cc",
         "./lib/ldbs.cc",
-        "./lib/logger.cc",
         "./lib/logrenderer.cc",
         "./lib/proto.cc",
         "./lib/readerwriter.cc",
@@ -180,7 +179,6 @@ cxxlibrary(
         "lib/imagewriter/imagewriter.h": "./lib/imagewriter/imagewriter.h",
         "lib/layout.h": "./lib/layout.h",
         "lib/ldbs.h": "./lib/ldbs.h",
-        "lib/logger.h": "./lib/logger.h",
         "lib/proto.h": "./lib/proto.h",
         "lib/readerwriter.h": "./lib/readerwriter.h",
         "lib/sector.h": "./lib/sector.h",
