+++ lib/core/logger.cc
@@ -1,9 +1,6 @@
 #include "lib/core/globals.h"
 #include "lib/core/bytes.h"
-#include "lib/fluxmap.h"
-#include "lib/sector.h"
-#include "lib/flux.h"
-#include "lib/logger.h"
+#include "lib/core/logger.h"
 
 static bool indented = false;
