+++ lib/fluxsource/dmkfluxsource.cc
@@ -4,7 +4,7 @@
 #include "lib/fluxsource/fluxsource.h"
 #include "lib/fluxsource/catweasel.h"
 #include "lib/proto.h"
-#include "lib/logger.h"
+#include "lib/core/logger.h"
 #include <fstream>
 #include <filesystem>
