+++ lib/fluxsource/hardwarefluxsource.cc
@@ -2,7 +2,7 @@
 #include "lib/config.h"
 #include "lib/flags.h"
 #include "lib/fluxmap.h"
-#include "lib/logger.h"
+#include "lib/core/logger.h"
 #include "lib/proto.h"
 #include "lib/usb/usb.h"
 #include "lib/fluxsource/fluxsource.h"
