+++ lib/fluxsource/scpfluxsource.cc
@@ -6,7 +6,7 @@
 #include "lib/fluxsource/fluxsource.h"
 #include "lib/scp.h"
 #include "lib/proto.h"
-#include "lib/logger.h"
+#include "lib/core/logger.h"
 #include <fstream>
 
 static int trackno(int strack)
