+++ lib/imagereader/dimimagereader.cc
@@ -4,7 +4,7 @@
 #include "lib/sector.h"
 #include "lib/imagereader/imagereader.h"
 #include "lib/image.h"
-#include "lib/logger.h"
+#include "lib/core/logger.h"
 #include "lib/proto.h"
 #include "lib/config.pb.h"
 #include <algorithm>
