+++ lib/imagereader/fdiimagereader.cc
@@ -5,7 +5,7 @@
 #include "lib/imagereader/imagereader.h"
 #include "lib/image.h"
 #include "lib/proto.h"
-#include "lib/logger.h"
+#include "lib/core/logger.h"
 #include "lib/config.pb.h"
 #include <algorithm>
 #include <iostream>
