+++ lib/imagereader/nfdimagereader.cc
@@ -4,7 +4,7 @@
 #include "lib/imagereader/imagereader.h"
 #include "lib/image.h"
 #include "lib/proto.h"
-#include "lib/logger.h"
+#include "lib/core/logger.h"
 #include "lib/config.pb.h"
 #include <algorithm>
 #include <iostream>
