+++ lib/imagewriter/d88imagewriter.cc
@@ -5,7 +5,7 @@
 #include "lib/image.h"
 #include "lib/config.pb.h"
 #include "lib/layout.h"
-#include "lib/logger.h"
+#include "lib/core/logger.h"
 #include <algorithm>
 #include <iostream>
 #include <fstream>
