+++ lib/imagewriter/diskcopyimagewriter.cc
@@ -4,7 +4,7 @@
 #include "lib/imagewriter/imagewriter.h"
 #include "lib/ldbs.h"
 #include "lib/image.h"
-#include "lib/logger.h"
+#include "lib/core/logger.h"
 #include "lib/config.pb.h"
 #include <algorithm>
 #include <iostream>
