+++ lib/imagewriter/nsiimagewriter.cc
@@ -4,7 +4,7 @@
 #include "lib/imagewriter/imagewriter.h"
 #include "lib/decoders/decoders.h"
 #include "lib/image.h"
-#include "lib/logger.h"
+#include "lib/core/logger.h"
 #include "arch/northstar/northstar.h"
 #include "lib/imagewriter/imagewriter.pb.h"
 #include <algorithm>
