+++ lib/logrenderer.cc
@@ -3,7 +3,7 @@
 #include "lib/fluxmap.h"
 #include "lib/sector.h"
 #include "lib/flux.h"
-#include "lib/logger.h"
+#include "lib/core/logger.h"
 
 namespace
 {
