+++ lib/vfs/amigaffs.cc
@@ -4,7 +4,7 @@
 #include "lib/config.pb.h"
 #include "lib/proto.h"
 #include "lib/layout.h"
-#include "lib/logger.h"
+#include "lib/core/logger.h"
 
 #include "adflib.h"
 #include "adf_blk.h"
